+++ typescript/tests.ts
@@ -88,6 +88,8 @@ RecoilRoot({
     set(writeableSelector, new DefaultValue());
   },
 });
+RecoilRoot({override: true});
+RecoilRoot({override: false});
 
 // Loadable
 function loadableTest(loadable: Loadable<number>) {