+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 
 - (Add new changes here as they land)
 - Improved TypeScript and Flow typing for `Loadable`s (#966)
+- Added override prop to RecoilRoot
 
 ## 0.2.0 (2021-3-18)